Delete restart directories for repeat runs
Repeat runs (`repeat: True` in `config.yaml`) are primarily for testing,
and there is no real need to preserve the restarts.  When I am doing
repeat runs, I usually do very many experiments (10s, if not 100s) and
keeping the restarts can consume huge storage.

This is a very crude patch which deletes any restarts produced by a
restart run.  It also determines the default run counter based on
`output` dirs if `restart` dirs do not exist.  It also resolves some
startup issues when counter is nonzero but restarts are absent.

In truth, this is all a bit bungled and need to be thought out a bit
better.  But for now it's fine.