Check if restarts are directories before deletion
Some users will add a symbolic link to the restart directory of another
run if they are forking an experiment.  shutil.rmtree() fails if the
target is not a directory, and will cause a crash.

We now check to see if the restart directory is a directory before
calling rmtree().

I decided not to use an exception because only directories will be
treated as acceptable for deletion.  Anything else will be treated as a
foreign non-payu file, and will be therefore be preserved.  (All bets
are off for non-payu directories).