Generalised MITgcm timing support
Turns out MITgcm supprts multiple ways to run the models, either with a
global timestep or independent momentum and tracer timesteps.

Similarly, duration can be set by either ntimesteps or by a start and
end time.

Some minimal support for these different time configurations are now
included.

This is just a quick patch, and is mainly to avoid crashing; a better
solution needs to be consistent with the source code.