Use cwd for collate jobname if unset in config
Collate jobs fail if there is no jobname set because the default config
path has not yet been specified.  This sets the default jobname to the
current working directory (cwd).

This is not a great solution, since it is done again elsewhere.  This is
just a temporary fix until this issue is properly sorted out.