API Token for repository creation
We now use a pre-generated github repository token to generate the
repository without authentication.

This token is currently stored in the .payu.yaml file.  Currently stored
locally, but probably ought to just be stored in $HOME.

The final solution is probably going to involve both SSH and HTTPS.

In any case, this is all successful with no user authentication.  The
downside is that there are now two non-password protected methods for
directly modifying (if not destroying) ones github accounts.  Need to
really make sure this is safe.