Github API token generation (keygen)
The `payu keygen` command now generates a github API token for use with
the runlog.  This enables passwordless pushing of model configurations
to github.

It also creates a plaintext copy of the github token in ones' home
directory, which is a potentially massive security risk.  Use at your
own risk.