payu keygen; payu push rework
All github configuration is now moved to the `payu keygen` function,
which is assumed to be run manually.  `payu push` now assumes that all
keys are already in place, and just pushes content to the repository.

I expect that `keygen` will be renamed at some point, it's more of a
"github setup" command (although a key is also generated).