FMS: Check return codes of mppnccombine pool
The following patch gathers the return codes for the pool of
mppnccombine jobs and checks their error codes.  If nonzero, then abort
the run.

(Not sure how this handles signals and other runtime crashes.)

This probably makes the checks inside cmdthread redundant, but need to
test it out.

Unfortunately it hasn't been tested.  This is partly as a demonstration
for potential merge.