collate: safer hyperthread disable
Memory eludes me on this, but it seems we wanted to disable
hyperthreading when doing collation.

The old test would simply look for any qsub_flags and remove any that
contain `hyperthread` (actually all of the letters but nevermind).  This
worked if it was a single word, e.g.

    qsub_flags: -lother=hyperthread

but failed when spaced, e.g.

    qsub_flags: -l other=hyperthread

This patch looks ahead after the `-l` flag and combines them into a
single argument.

Still quite imperfect, but a safer solution.