Exclude python script from PYTHONPATH
Construction of PYTHONPATH from `sys.path` was including the directory
of the executing script, which was unnecessary since we are constructing
PYTHONPATH for the next script, not the current one (which would likely
have the same path anyway).  This was also creating useless double
entries.

We resolve this by skipping the first element.