PYTHONPATH cleanup; install cleanup
Instead of trying to intuit the PYTHONPATH from `sys.path` and guessing
at the default paths to remove, we just manually pass along the existing
PYTHONPATH.

The old way probably never made any sense; the idea was to try and
determine PYTHONPATH from available paths, and then remove any default
paths.  But this would probably just end up with the contents of
PYTHONPATH anyway.

We have also trimmed out a few unnecessary elements of setup.py and
setup.cfg (with more to come).