User restart input; repeat run support and fix
This patch adds support for a user-directed restart directory.
Directory is provided by the `restart: $path` argument in config.yaml.
This directory is only used for counter 0; otherwise, we just use the
latest restart directory.

Currently only supports absolute paths, although this could be changed.

This patch also improves suport for repeated runs (self.repeat_run, set
by `repeat: True` in config.yaml).  We now use repeat_run to set the
prior_restart_path of the experiment and model, which allows us to
either target this as the prior restart, the user restart, or None for
a cold start.

This change to repeat_run also allows to avoid references to repeat_run
in the drivers.  Some of these have been fixed (mom, cice, and the
generic Model), but they are still present in other models and probably
don't work yet.  We'll fix these on a case-by-case basis.