Bugfix: ncpus can be omitted if set in submodels
The `ncpus` config of config.yaml can now be removed if the cpus are set
for each submodel; the total cpus is inferred from the sum of submodel
cpus.

This was implemented previously, but was broken when submodels were
changed from a dict to a list.  This patch fixes the bug.