MOM: Mask table file search
We have included an option for MOM models to search for a mask_table
file suitable for the run.  This is very limited in scope, and will only
work if there is a file of format `mask_table.NMASK.MxN` where NMASK is
the number of masked CPUs and MxN is the nominal CPU layout.
In particular, the CPU request must match `M*N - NMASK` and the `layout`
namelist variable must be equal to (M, N).

This may only have limited use, and may not persist for much longer, but
is useful for dynamically selecting a mask table against a precomputed
set of tables.