Added git commit id manipulation routines to fsops.
Made a git_id routine to store and retrieve git commit id/hash from
PayuManifest objects.

Now check if git hash in input manifest is an ancestor of the last
git commit for config.yaml and emit a warning message about deleting
the input manifest if the input directories need to be re-scanned.