Removal of MOM5 core2iaf support
The `core2iaf` function was used to process the interannual CORE
forcings across multiple MOM runs.  Given that the community has moved
on to JRA forcing, and that promised performance improvement of core2iaf
has not been an issue for years, and that active or published
experiments ever used this function, it should be safe to remove it.