Proto-scheduler class, pre-support for slurm
This is a very minimal implementation of a Scheduler class, in order to
handle scheduler-specific operations.

It includes a basic Factory for controlling the Scheduler instantiation,
and a single method for returning the submission command line.

I do not expect things to keep working like this, it is only meant to be
the most minimal implementation to support both PBS and Slurm.

Hopefully more changes to come.