Flag a MOM6 run as a restart if using prior restart path (#324)
Using the mechanism to point to an external restart directory:

  restart: /path/to/restart

doesn't modify the experiment counter, because as far as payu is
concerned, this is the first run. This breaks the logic the MOM6
module was using to determine whether to flag a run as a restart or
not. In the case that the prior restart path has been set, we want to
explicitly perform a restart run.