Wrap ldd in try/except as executables on pawsey seem to be statically linked. Also for the same reason don't assume LD_LIBRARY_PATH is set.
Commented out call to load_modules. Pawsey has a lot of default modules
that it relies on, so can't reliably monkey with that.

Removed a couple of the bespoke flags Marshall added to the slurm
scheduler, and also explicitly pass through the PAYU environment
variables. Also set in the current environment, but that didn't seem
to make it through to the submitted job.