Improved config log and restart handling
Now only build the config log during setup, as I learned that expt.n_runs can be set to control resubmission rather than setting the environment variable directly. During archiving, any restarts from the previous stage not updated by the current stage are copied to the next set of restarts, so all that is required for portable restarting is a single restart directory.